Added type property to the token "instances" for easier debugging.
fixes #542
