+++ test/parse/grammar/checks_spec.ts
@@ -969,29 +969,24 @@ describe("The no non-empty lookahead validation", () => {
         )
     })
 
-    it(
-        "will throw an error when there are too many alternatives in an alternation",
-        () => {
-            const alternatives = []
-            for (let i = 0; i < 256; i++) {
-                alternatives.push(
-                    new Flat([new NonTerminal("dummyRule", dummyRule)])
-                )
-            }
-
-            const ruleWithTooManyAlts = new Rule("blah", [
-                new Alternation(alternatives)
-            ])
-
-            const actual = validateTooManyAlts(ruleWithTooManyAlts)
-            expect(actual).to.have.lengthOf(1)
-            expect(actual[0].type).to.equal(
-                ParserDefinitionErrorType.TOO_MANY_ALTS
-            )
-            expect(actual[0].ruleName).to.equal("blah")
-            expect(actual[0].message).to.contain(
-                "An Alternation cannot have more than 256 alternatives"
+    it("will throw an error when there are too many alternatives in an alternation", () => {
+        const alternatives = []
+        for (let i = 0; i < 256; i++) {
+            alternatives.push(
+                new Flat([new NonTerminal("dummyRule", dummyRule)])
             )
         }
-    )
+
+        const ruleWithTooManyAlts = new Rule("blah", [
+            new Alternation(alternatives)
+        ])
+
+        const actual = validateTooManyAlts(ruleWithTooManyAlts)
+        expect(actual).to.have.lengthOf(1)
+        expect(actual[0].type).to.equal(ParserDefinitionErrorType.TOO_MANY_ALTS)
+        expect(actual[0].ruleName).to.equal("blah")
+        expect(actual[0].message).to.contain(
+            "An Alternation cannot have more than 256 alternatives"
+        )
+    })
 })
